Separate works-specific and api-wide tests, remove the "V2" distinction
