+++ api/api/src/test/scala/uk/ac/wellcome/platform/api/ApiSwaggerTest.scala
@@ -6,9 +6,9 @@ import akka.http.scaladsl.model.ContentTypes
 import io.circe.Json
 import uk.ac.wellcome.platform.api.models.SearchQueryType
 import uk.ac.wellcome.platform.api.rest.{MultipleWorksParams, SingleWorkParams}
-import uk.ac.wellcome.platform.api.works.v2.ApiV2WorksTestBase
+import uk.ac.wellcome.platform.api.works.ApiWorksTestBase
 
-class ApiSwaggerTest extends ApiV2WorksTestBase with Matchers {
+class ApiSwaggerTest extends ApiWorksTestBase with Matchers {
   val worksEndoint = "/works"
   val workEndpoint = "/works/{id}"
   it("should return valid json object") {
